requirement.txt.txt
streamlit
Pillow
torch==2.1.0
torchvision==0.16.0
onnx==1.14.1
onnxruntime==1.16.3
tensorflow==2.15.0
tf2onnx==1.16.1
numpy==1.26.2
onnx-tf==1.10.0
tensorflow_probability==0.23.0
ipykernel



